+++ expo/gbe/forms/stage_info_form.py
@@ -7,6 +7,7 @@ from gbe_forms_text import (
     act_help_texts,
     prop_labels,
 )
+from django.core.exceptions import ValidationError
 
 
 class StageInfoForm(ModelForm):
